Refactor: refactoring get refresh token & retry
