+++ front/src/pages/ModifyPostView.jsx
@@ -3,6 +3,7 @@ import axios from 'axios';
 import styled from 'styled-components';
 import { useParams } from 'react-router-dom';
 import useFetchData from '../ components/hooks/getPostList';
+import { refreshTokenAndRetry } from '../api/user';
 
 export default function ModifyPostView() {
   const { board_id } = useParams();
@@ -199,34 +200,19 @@ export default function ModifyPostView() {
       console.log(error.response.data);
       if (error.response.data.msg == '인증이 실패했습니다.') {
         console.log(error.response.data.msg);
-        const refreshData = {
-          refreshToken: localStorage.getItem('refreshToken'),
-        };
+
         try {
-          const refreshResponse = await axios.post(
-            'http://1.246.104.170:8080/sign/reissue',
-            refreshData
-          );
-
-          // 새로운 액세스 토큰 저장
-          const newAccessToken = refreshResponse.data.accessToken;
-          const newRefreshToken = refreshResponse.data.refreshToken;
-          localStorage.setItem('token', newAccessToken);
-          localStorage.setItem('refreshToken', newRefreshToken);
-
-          // 새로운 액세스 토큰을 사용하여 원래의 요청 다시 보내기
-          const retryResponse = await axios.put(
+          const retryResponse = await refreshTokenAndRetry(
+            'put',
             `http://1.246.104.170:8080/boards/${board_id}`,
             modifyData,
             {
-              headers: {
-                'X-AUTH-TOKEN': newAccessToken,
-              },
+              'X-AUTH-TOKEN': localStorage.getItem('token'),
             }
           );
-          console.log('글 수정 성공 (재시도)');
-        } catch (refreshError) {
-          console.error('새로운 액세스 토큰 얻기 실패', refreshError);
+          console.log(retryResponse);
+        } catch (retryError) {
+          console.log(retryError);
         }
       }
     }
